Add object_property_get_child().
Adding a direct accessor to a child property.

In the existing implementation, object_property_get() must be used,
with with a visitor, implementing the 'type_str' callback,
receiving the child's canonical path.

In the new implementation, the child is returned directly.
For link properties, object_property_get_link() is used
to resolve the link.

Also, in the new implementation, object_property_get_child() is used
as a subroutine of object_resolve_abs_path(). This changes the way
object_resolve_abs_path() operates, moving away from directly peeking
the property's 'opaque' field to using object_property_get_link().

Thus, in the mew implementation link properties are resolved in the same way,
as they are when an absolute path is resolved.

Errors relevant to the operation, QERR_OBJECT_PROPERTY_NOT_FOUND
and QERR_OBJECT_PROPERTY_INVALID_TYPE were added.

Also, in the new implementation, some common sense refactoring was done
in the file 'qom/object.c' in the code extracting child and link properties.

Signed-off-by: Alexander Barabash <alexander_barabash@mentor.com>